.6.0-avaudio: Auto stash before rebase of "cocos/v3.6.0"
