+++ native/CMakeLists.txt
@@ -345,14 +345,10 @@ if(USE_AUDIO)
     elseif(APPLE)
         cocos_source_files(
                                     cocos/audio/apple/AudioDecoder.h
-            NO_WERROR   NO_UBUILD   cocos/audio/apple/AudioPlayer.mm
             NO_WERROR   NO_UBUILD   cocos/audio/apple/AudioDecoder.mm
-                                    cocos/audio/apple/AudioPlayer.h
-            NO_WERROR   NO_UBUILD   cocos/audio/apple/AudioEngine-newImpl.mm
+            NO_WERROR   NO_UBUILD   cocos/audio/apple/AudioEngine-Impl.mm
                                     cocos/audio/apple/AudioMacros.h
-            NO_WERROR   NO_UBUILD   cocos/audio/apple/AudioCache.mm
-                                    cocos/audio/apple/AudioCache.h
-                                    cocos/audio/apple/AudioEngine-newImpl.h
+                                    cocos/audio/apple/AudioEngine-Impl.h
         )
     endif()
